SRCH-3973: added components
