+++ app/javascript/components/Alert/Alert.css
@@ -1,0 +1,3 @@
+.alert-wrapper{
+  margin-bottom: 3rem;
+}
