+++ app/javascript/components/Results/Results.css
@@ -244,6 +244,7 @@
 .comment-period-ended{
   color: #71767A;
   line-height: 162%;
+  margin-top: 0.5rem;
 }
 
 .more-fed-register-link{
