+++ app/javascript/components/SearchBar/SearchBar.css
@@ -1,7 +1,3 @@
 .usa-input{
   max-width: none;
 }
-
-.alert-wrapper{
-  margin-bottom: 3rem;
-}
